sanitizing chat messages in the chat message model
