+++ js/views/chat/Conversation.js
@@ -269,28 +269,12 @@ export default class extends baseVw {
       throw new Error('Please provide a message to send.');
     }
 
-    let message = msg;
     this.lastTypingSentAt = null;
-
-    // Convert any emoji placeholder (e.g :smiling_face:) into
-    // emoji unicode characters.
-    const emojiPlaceholderRegEx = new RegExp(':.+?:', 'g');
-    const matches = message.match(emojiPlaceholderRegEx, 'g');
-
-    if (matches) {
-      matches.forEach(match => {
-        const emoji = getEmojiByName(match);
-
-        if (emoji && emoji.char) {
-          message = message.replace(match, emoji.char);
-        }
-      });
-    }
 
     const chatMessage = new ChatMessage({
       peerId: this.guid,
       subject: this.subject,
-      message,
+      message: msg,
     });
 
     const save = chatMessage.save();
